refactor: cleanup types and consts
