--- packages/apps/reputation-oracle/server/src/common/interfaces/index.ts
@@ -1,2 +1,0 @@
-export * from './base';
-export * from './user';
