--- packages/apps/reputation-oracle/server/src/common/types/index.ts
@@ -1,1 +1,0 @@
-export * from './request';
