+++ packages/apps/reputation-oracle/server/src/modules/storage/storage.service.ts
@@ -11,7 +11,7 @@ import * as Minio from 'minio';
 import crypto from 'crypto';
 import { UploadedFile } from '../../common/interfaces/s3';
 import { Web3Service } from '../web3/web3.service';
-import { FortuneFinalResult } from '../../common/dto/result';
+import { FortuneFinalResult } from '../../common/interfaces/job-result';
 import { S3ConfigService } from '../../common/config/s3-config.service';
 import { PGPConfigService } from '../../common/config/pgp-config.service';
 import { isNotFoundError } from '../../common/utils/minio';
