+++ packages/apps/reputation-oracle/server/src/modules/user/user.controller.ts
@@ -29,7 +29,7 @@ import {
 } from './user.dto';
 import { JwtAuthGuard } from '../../common/guards';
 import { HCaptchaGuard } from '../../common/guards/hcaptcha';
-import { RequestWithUser } from '../../common/types';
+import { RequestWithUser } from '../../common/interfaces/request';
 import { prepareSignatureBody } from '../../utils/web3';
 import { UserService } from './user.service';
 import { Public } from '../../common/decorators';
